fix eagle mediawiki integration bug
Code was crashing because of a missing variable.

In `4dc207fc6f4c80f32372ca334736ef9252ad6b28`, code was copied and
pasted from another section and used a variable called`workUrn`. That
variable was not defined in the context it was used.

In addition, the source that was copied and pasted had a bug in it,
using `gsub` instead of `gsub!`. That bug was introduced in
`6d1ef15039ab6aea9588ec79936009de81de4e41`. This commit fixes both
issues and is another step towards fixing EAGLE integration.